Github Issues Fetcher v0.0.1 Issues.TableFormatter

Functions

column_widths_for(rows)

print_row(row, format)

print_table(source_items, list)

printable(str)

row_format_for(column_widths)

Return a format string that hard codes the widths of a set of columns. We put " | " between each column.

Example

iex> import Issues.TableFormatter
iex> column_widths = [2, 3, 4]
iex> row_format_for(column_widths)
"~-2s | ~-3s | ~-4s~n"

separator_for(column_widths)

Generate the line that goes below the column headings. It is a string of hyphens, with + signs where the vertical bar between the columns goes.

Example

iex> import Issues.TableFormatter
iex> column_widths = [2, 3, 4]
iex> separator_for(column_widths)
"---+-----+-----"

source_items_to_rows(source_items, fields)